Create experimental.bisect module, put bisect_left in that instead of in sigma.py
