+++ software/contrib/sigma.py
@@ -13,47 +13,10 @@ from europi_script import EuroPiScript
 import configuration
 import time
 
+from experimental.bisect import bisect_left
 from experimental.knobs import *
 from experimental.random_extras import normal
 from experimental.screensaver import Screensaver
-
-
-def bisect_left(a, x, lo=0, hi=None, *, key=None):
-    """Return the index where to insert item x in list a, assuming a is sorted.
-
-    The return value i is such that all e in a[:i] have e < x, and all e in
-    a[i:] have e >= x.  So if x already appears in the list, a.insert(i, x) will
-    insert just before the leftmost x already there.
-
-    Optional args lo (default 0) and hi (default len(a)) bound the
-    slice of a to be searched.
-
-    A custom key function can be supplied to customize the sort order.
-
-    Copied from https://github.com/python/cpython/blob/3.12/Lib/bisect.py
-    """
-
-    if lo < 0:
-        raise ValueError('lo must be non-negative')
-    if hi is None:
-        hi = len(a)
-    # Note, the comparison uses "<" to match the
-    # __lt__() logic in list.sort() and in heapq.
-    if key is None:
-        while lo < hi:
-            mid = (lo + hi) // 2
-            if a[mid] < x:
-                lo = mid + 1
-            else:
-                hi = mid
-    else:
-        while lo < hi:
-            mid = (lo + hi) // 2
-            if key(a[mid]) < x:
-                lo = mid + 1
-            else:
-                hi = mid
-    return lo
 
 
 class OutputBin:
